refactor(orchestrator): trim to interrupt-proof teardown; drop preservation-prune
Preservation-path pruning is covered downstream by coder_eval_uipath #57 and
isn't reachable on current configs; keep this PR to the teardown fix.
